requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
chromadb>=0.5.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
